@@ -4,53 +4,53 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "build": "nx build",
     "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/forms": "~13.1.0",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
-    "@nrwl/angular": "13.5.2",
+    "@angular/animations": "14.0.4",
+    "@angular/common": "14.0.4",
+    "@angular/compiler": "14.0.4",
+    "@angular/core": "14.0.4",
+    "@angular/forms": "14.0.4",
+    "@angular/platform-browser": "14.0.4",
+    "@angular/platform-browser-dynamic": "14.0.4",
+    "@angular/router": "14.0.4",
+    "@nrwl/angular": "14.3.6",
     "rxjs": "~7.4.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.0",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.1.0",
-    "@angular/compiler-cli": "~13.1.0",
-    "@angular/language-service": "~13.1.0",
-    "@nrwl/cli": "13.5.2",
-    "@nrwl/cypress": "13.5.2",
-    "@nrwl/eslint-plugin-nx": "13.5.2",
-    "@nrwl/jest": "13.5.2",
-    "@nrwl/linter": "13.5.2",
-    "@nrwl/tao": "13.5.2",
-    "@nrwl/workspace": "13.5.2",
-    "@types/jest": "27.0.2",
+    "@angular-devkit/build-angular": "14.0.4",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.4",
+    "@angular/language-service": "14.0.4",
+    "@nrwl/cli": "14.3.6",
+    "@nrwl/cypress": "14.3.6",
+    "@nrwl/eslint-plugin-nx": "14.3.6",
+    "@nrwl/jest": "14.3.6",
+    "@nrwl/linter": "14.3.6",
+    "@nrwl/workspace": "14.3.6",
+    "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "cypress": "^9.1.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "prettier": "^2.5.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.4.3"
+    "jest": "27.5.1",
+    "jest-preset-angular": "11.1.2",
+    "prettier": "2.6.2",
+    "ts-jest": "27.1.4",
+    "typescript": "4.7.4",
+    "nx": "14.3.6"
   }
 }
